cleaning up code moved functions into new .c files fixed small graphical glitch on title screen
