+++ src/main.c
@@ -89,42 +89,7 @@ void	nsew_textures(t_game *game)
 	if (west) 
 		game->textures[TEX_WALLW] = texture_load(game->app.mlx, west->content);
 }
-/*
-void	title_imgs_load_all(t_game *game)
-{
-	game->textures[TEX_TITLE_MISSION_COMPLETED] = texture_load(game->app.mlx, "assets/menu/mission_competed.xpm");
-	game->textures[TEX_TITLE_MISSION_FAILED] = texture_load(game->app.mlx, "assets/menu/mission_failed.xpm");
-	game->textures[TEX_TITLE_NEXT_LEVEL_0] = texture_load(game->app.mlx, "assets/menu/next_level_0.xpm");
-	game->textures[TEX_TITLE_NEXT_LEVEL_1] = texture_load(game->app.mlx, "assets/menu/next_level_1.xpm");
-	game->textures[TEX_TITLE_RETRY_0] = texture_load(game->app.mlx, "assets/menu/retry_0.xpm");
-	game->textures[TEX_TITLE_RETRY_1] = texture_load(game->app.mlx, "assets/menu/retry_1.xpm");
-	game->textures[TEX_TITLE_CREATED_CHRIS_KYLE] = texture_load(game->app.mlx, "assets/menu/created_chris_kyle.xpm");
-	game->textures[TEX_TITLE_MAIN_MENU_0] = texture_load(game->app.mlx, "assets/menu/main_menu_0.xpm");
-	game->textures[TEX_TITLE_MAIN_MENU_1] = texture_load(game->app.mlx, "assets/menu/main_menu_1.xpm");
-}
 
-
-void	generate_all_textures(t_game *game)
-{
-	enemy_load_directory(game, "assets/enemy_sprites/");
-	game->textures[TEX_WINDOW] = texture_load(game->app.mlx, "assets/window.xpm");
-	game->textures[TEX_DOOR] = game->textures[TEX_WINDOW];
-	game->textures[TEX_FLOOR] = texture_load(game->app.mlx, "assets/metal_walkway_acg.xpm");
-	game->textures[TEX_CEILING] = game->textures[TEX_FLOOR];
-	game->textures[TEX_FIVE_LIGHTS] = texture_load(game->app.mlx, "assets/five_lights_wall.xpm");
-	game->textures[TEX_UI_CONTROLS] = texture_load(game->app.mlx, "assets/ui/controls.xpm");
-	game->textures[TEX_UI_INTERACT_BRIGHT] = texture_load(game->app.mlx, "assets/ui/interact_bright.xpm");
-	game->textures[TEX_UI_INTERACT_DUL] = texture_load(game->app.mlx, "assets/ui/interact_dal.xpm");
-	game->textures[TEX_UI_DOOR_LOCKED] = texture_load(game->app.mlx, "assets/ui/promt_door_locked.xpm");
-	game->textures[TEX_UI_TASK_INACTIVE] = texture_load(game->app.mlx, "assets/ui/inactive_task.xpm");
-	game->loaded_index[0] = TEX_ARRAY_SIZE;
-	title_imgs_load_all(game);
-	nsew_textures(game);
-
-//	printf("Loading: %s\n", north->content);
-}
-
-*/
 void generate_textures(t_game *game)
 {
 	game->textures[TEX_SKYBOX] = texture_load(game->app.mlx, "assets/skybox.xpm");
@@ -155,8 +120,6 @@ void	setup_world(int argc, char **argv, t_game *game)
 	nsew_textures(game);
 }
 
-
-
 int	main(int argc, char **argv)
 {
 	t_game	game;
@@ -165,7 +128,7 @@ int	main(int argc, char **argv)
 	game.world = malloc(sizeof(t_world));
 	*game.world = (t_world){0};
 	game.app.mlx = mlx_init();
-	title_screen(&game);
+	title_setup_screens(&game);
 	game.loaded_index[0] = 0;
 	game.loaded_index[1] = 1;
 	game.loaded_index[2] = 0;
@@ -178,10 +141,7 @@ int	main(int argc, char **argv)
 	game.display_ui = false;
 	generate_textures(&game);
 	if (argc == 2)
-	{
-	//	generate_all_textures(&game);
 		setup_world(argc, argv, &game);
-	}
 	game.rt1 = texture_create(game.app.mlx, R_WIDTH, R_WIDTH);
 	game.rt0 = texture_create(game.app.mlx, SCR_WIDTH, SCR_HEIGHT);
 	game.rt2 = texture_create(game.app.mlx, SCR_WIDTH, SCR_HEIGHT);
